refactor : 파일 구분 깔끔하게
