+++ src/main/java/com/opus/opus/modules/contest/api/ContestController.java
@@ -13,7 +13,7 @@ import com.opus.opus.modules.contest.application.dto.response.ContestResponse;
 import com.opus.opus.modules.contest.application.dto.response.ContestVotesLimitResponse;
 import com.opus.opus.modules.contest.application.dto.response.VotePeriodResponse;
 import com.opus.opus.modules.member.domain.Member;
-import com.opus.opus.modules.team.application.TeamVoteQueryService;
+import com.opus.opus.modules.team.application.TeamQueryService;
 import com.opus.opus.modules.team.application.dto.ImageResponse;
 import com.opus.opus.modules.team.application.dto.response.MemberVoteCountResponse;
 import jakarta.validation.Valid;
@@ -44,7 +44,7 @@ public class ContestController {
 
     private final ContestCommandService contestCommandService;
     private final ContestQueryService contestQueryService;
-    private final TeamVoteQueryService teamVoteCommandService;
+    private final TeamQueryService teamQueryService;
 
     @GetMapping("/{contestId}/image/banner")
     public ResponseEntity<Resource> getContestBanner(@PathVariable final Long contestId) {
@@ -144,7 +144,7 @@ public class ContestController {
     @GetMapping("/{contestId}/votes/me")
     public ResponseEntity<MemberVoteCountResponse> getMemberVoteCount(@PathVariable Long contestId,
                                                                       @LoginMember Member member) {
-        MemberVoteCountResponse response = teamVoteCommandService.getMemberVoteCount(member.getId(), contestId);
+        MemberVoteCountResponse response = teamQueryService.getMemberVoteCount(member.getId(), contestId);
         return ResponseEntity.ok(response);
     }
 }
