--- src/test/java/com/opus/opus/team/application/TeamVoteQueryServiceTest.java
@@ -1,85 +1,0 @@
-package com.opus.opus.team.application;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.opus.opus.contest.ContestFixture;
-import com.opus.opus.helper.IntegrationTest;
-import com.opus.opus.member.MemberFixture;
-import com.opus.opus.modules.contest.domain.Contest;
-import com.opus.opus.modules.contest.domain.dao.ContestRepository;
-import com.opus.opus.modules.member.domain.Member;
-import com.opus.opus.modules.member.domain.dao.MemberRepository;
-import com.opus.opus.modules.team.application.TeamVoteQueryService;
-import com.opus.opus.modules.team.application.dto.response.MemberVoteCountResponse;
-import com.opus.opus.modules.team.domain.Team;
-import com.opus.opus.modules.team.domain.dao.TeamRepository;
-import com.opus.opus.modules.team.domain.dao.TeamVoteRepository;
-import com.opus.opus.team.TeamFixture;
-import com.opus.opus.team.TeamVoteFixture;
-import java.time.LocalDateTime;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-public class TeamVoteQueryServiceTest extends IntegrationTest {
-
-    @Autowired
-    private TeamVoteQueryService teamVoteQueryService;
-
-    @Autowired
-    private TeamRepository teamRepository;
-    @Autowired
-    private MemberRepository memberRepository;
-    @Autowired
-    private TeamVoteRepository teamVoteRepository;
-    @Autowired
-    private ContestRepository contestRepository;
-
-    private Contest contest;
-    private Team team;
-    private Member member;
-
-    @BeforeEach
-    void setUp() {
-        Contest newContest = ContestFixture.createContest();
-        newContest.updateVotePeriod(LocalDateTime.now().minusDays(1), LocalDateTime.now().plusDays(1));
-        newContest.updateMaxVotesLimit(2);
-        contest = contestRepository.save(newContest);
-
-        team = teamRepository.save(TeamFixture.createTeamWithContestId(contest.getId()));
-        member = memberRepository.save(MemberFixture.createMember());
-    }
-
-    @Test
-    @DisplayName("[성공] 사용자의 남은 투표 개수를 조회할 수 있다.")
-    void 사용자의_남은_투표_개수를_조회할_수_있다() {
-        teamVoteRepository.save(TeamVoteFixture.createTeamVote(team, member.getId(), true));
-
-        MemberVoteCountResponse response = teamVoteQueryService.getMemberVoteCount(member.getId(), contest.getId());
-
-        assertThat(response.remainingVotesCount()).isEqualTo(1L);
-        assertThat(response.maxVotesLimit()).isEqualTo(2L);
-    }
-
-    @Test
-    @DisplayName("[성공] 투표하지 않은 사용자는 최대 투표 수만큼 남은 투표 개수가 있다.")
-    void 투표하지_않은_사용자는_최대_투표_수만큼_남은_투표_개수가_있다() {
-        MemberVoteCountResponse response = teamVoteQueryService.getMemberVoteCount(member.getId(), contest.getId());
-
-        assertThat(response.remainingVotesCount()).isEqualTo(2L);
-        assertThat(response.maxVotesLimit()).isEqualTo(2L);
-    }
-
-    @Test
-    @DisplayName("[성공] 취소한 투표는 카운트에서 제외된다.")
-    void 취소한_투표는_카운트에서_제외된다() {
-        teamVoteRepository.save(TeamVoteFixture.createTeamVote(team, member.getId(), true));
-        Team secondTeam = teamRepository.save(TeamFixture.createTeamWithContestId(contest.getId()));
-        teamVoteRepository.save(TeamVoteFixture.createTeamVote(secondTeam, member.getId(), false));
-
-        MemberVoteCountResponse response = teamVoteQueryService.getMemberVoteCount(member.getId(), contest.getId());
-
-        assertThat(response.remainingVotesCount()).isEqualTo(1L);
-    }
-}
